requirements.txt
transformers
transformers[sentencepiece]
torch
matplotlib
numpy
datasets
sacrebleu
rouge_score
pandas
py7zr
pyYAML
nltk
tqdm
notebook
boto3
mypy-boto3-s3
python-box==6.0.2
fastapi==0.95.2
uvicorn==0.22.0
Jinja2==3.1.2
ensure